"Media Player - added sliders"

+++ src/sample/ManageCreationController.java
@@ -1,18 +1,25 @@
 package sample;
 
+import javafx.beans.InvalidationListener;
+import javafx.beans.Observable;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
+import javafx.scene.input.MouseDragEvent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.media.Media;
 import javafx.scene.media.MediaPlayer;
 import javafx.scene.media.MediaView;
 import javafx.stage.Stage;
+import javafx.util.Duration;
 
 import java.io.File;
 import java.io.IOException;
@@ -22,6 +29,8 @@ import java.util.Optional;
 
 public class ManageCreationController {
 
+    @FXML private Slider _mediaSlider;
+    @FXML private Slider _volumeSlider;
     @FXML private MediaView _mediaView;
     private ObservableList<String> _items;
     private File _directory;
@@ -30,7 +39,6 @@ public class ManageCreationController {
     @FXML private Button _playCreationButton;
     @FXML private Button _playButton;
     @FXML private Button _pauseButton;
-    @FXML private Button _restartButton;
     @FXML private Button _fastPlayButton;
     @FXML private Button _slowPlayButton;
     @FXML private ListView _creationList;
@@ -60,28 +68,36 @@ public class ManageCreationController {
 
         if (validSelection()) {
 
-            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-            alert.setTitle("Confirmation");
-            alert.setHeaderText("You have chosen " + _creationChosen);
-            alert.setContentText("Are you sure you want to delete " + _creationChosen);
-            Optional<ButtonType> result = alert.showAndWait();
-            String cmd = "rm -rf ./Files/creations/" + _creationChosen + "; rm -rf ./Files/keywords/" + _creationChosen;
-            if (result.get() == ButtonType.OK) {
-                {
-                    try {
-                        Process process = new ProcessBuilder("/bin/bash", "-c", cmd).start();
-                        process.waitFor();
-                    } catch (IOException | InterruptedException e) {
+            if (player != null) {
+                Alert alert = new Alert(Alert.AlertType.ERROR);
+                alert.setTitle("Error");
+                alert.setHeaderText("Middle of Process!");
+                alert.setContentText("Please Wait for the Video to Finish");
+                alert.showAndWait();
+            } else {
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("Confirmation");
+                alert.setHeaderText("You have chosen " + _creationChosen);
+                alert.setContentText("Are you sure you want to delete " + _creationChosen);
+                Optional<ButtonType> result = alert.showAndWait();
+                String cmd = "rm -rf ./Files/creations/" + _creationChosen + "; rm -rf ./Files/keywords/" + _creationChosen;
+                if (result.get() == ButtonType.OK) {
+                    {
+                        try {
+                            Process process = new ProcessBuilder("/bin/bash", "-c", cmd).start();
+                            process.waitFor();
+                        } catch (IOException | InterruptedException e) {
+                        }
+                        Alert alert2 = new Alert(Alert.AlertType.INFORMATION);
+                        alert2.setTitle("Information Dialog");
+                        alert2.setHeaderText("Success!");
+                        alert2.setContentText("You have deleted " + _creationChosen);
+                        alert2.showAndWait();
+
+                        _creationChosen = null;
+                        refresh();
+
                     }
-                    Alert alert2 = new Alert(Alert.AlertType.INFORMATION);
-                    alert2.setTitle("Information Dialog");
-                    alert2.setHeaderText("Success!");
-                    alert2.setContentText("You have deleted " + _creationChosen);
-                    alert2.showAndWait();
-
-                    _creationChosen = null;
-                    refresh();
-
                 }
             }
         }
@@ -98,18 +114,55 @@ public class ManageCreationController {
                 player.stop();
                 player.dispose();
             }
+
             player = new MediaPlayer(vid);
 
-            player.setAutoPlay(true);
             _mediaView.setMediaPlayer(player);
+
+            player.setAutoPlay(true);
+
+
             player.setOnReady(() -> {
+                sliderSetUp();
+                _mediaSlider.setOpacity(1.0);
             });
             player.setOnEndOfMedia(() -> {
                 player.dispose();
                 _mediaView.setMediaPlayer(null);
                 player=null;
+                _mediaSlider.setOpacity(0.0);
             });
         }
+    }
+
+    private void sliderSetUp() {
+
+        _volumeSlider.setValue(player.getVolume() * 100);
+        _volumeSlider.valueProperty().addListener(new InvalidationListener() {
+            @Override
+            public void invalidated(Observable observable) {
+                player.setVolume(_volumeSlider.getValue()/100);
+            }
+        });
+
+        player.currentTimeProperty().addListener(new ChangeListener<Duration>() {
+            @Override
+            public void changed(ObservableValue<? extends Duration> observable, Duration oldValue, Duration newValue) {
+                _mediaSlider.setValue(newValue.toMillis());
+
+            }
+        });
+        _mediaSlider.setOnMousePressed(new EventHandler<MouseEvent>() {
+            @Override
+            public void handle(MouseEvent event) {
+                player.seek(Duration.millis(_mediaSlider.getValue()));
+            }
+        });
+
+
+
+        _mediaSlider.setMax(player.getTotalDuration().toMillis());
+        _mediaSlider.setValue(0);
     }
 
     @FXML public void playMediaOnClick(ActionEvent actionEvent) {
@@ -171,8 +224,8 @@ public class ManageCreationController {
         return true;
     }
 
-    private boolean mediaExist(){
-        if(player!=null){
+    private boolean mediaExist() {
+        if (player != null) {
             return true;
         }
         return false;
